backend/requirements.txt
# Invoice Processing & Fraud Detection — Backend
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
opencv-python-headless==4.10.0.84
numpy==2.1.3
pandas==3.0.2
pillow==11.0.0
pytesseract==0.3.13
scikit-learn==1.5.2
joblib==1.4.2
rapidfuzz==3.10.1
aiofiles==24.1.0
spacy==3.8.7
